+++ src/common/CMakeLists.txt
@@ -35,7 +35,6 @@ set(common_sources
   download.cpp
   util.cpp
   i18n.cpp
-  memwipe.c
   password.cpp
   perf_timer.cpp
   threadpool.cpp
@@ -64,7 +63,6 @@ set(common_private_headers
   util.h
   varint.h
   i18n.h
-  memwipe.h
   password.h
   perf_timer.h
   stack_trace.h
@@ -92,9 +90,6 @@ target_link_libraries(common
     ${OPENSSL_LIBRARIES}
     ${EXTRA_LIBRARIES})
 
-if(HAVE_C11)
-SET_PROPERTY(SOURCE memwipe.c PROPERTY COMPILE_FLAGS -std=c11)
-endif()
 
 #aeon_install_headers(common
 #  ${common_headers})
